Extract versions correctly, simplify
The code to extract module versions wasn't a bit
old and missed some cases, so I fixed that.

I simplified the code to check for installed 
module advisories because it was getting a bit 
mind bending and the something is weird with
version.pm.

This now includes more support for tracking perl
advisories too.