New `--exclude-file` option, to exclude reports from a file.
While `--exclude` can be provided multiple times on the command line, it
would be nice to have the ability to not only exclude multiple reports
at once, but to also be able to document _why_ those exclusions are
being performed.

Thus, allow for an `--exclude-file` to be provided.  One entry per line,
and which can be documented with `#` style comments.