New `--exclude-file` option, to exclude/ignore reports from file (#15)
* Fix POD; this is `exclude` (singular, not plural).

Realigns the POD with the actual name of the parameter being used.

* Switch from `Catalyst` to `CPAN` in tests; has >1 advisory.

Switching to a distribution that has more than one advisory makes it
easier to detect that we are properly excluding advisories (as the one
we are excluding would no longer be present in the output, but the other
remaining one would).

* New `--exclude-file` option, to exclude reports from a file.

While `--exclude` can be provided multiple times on the command line, it
would be nice to have the ability to not only exclude multiple reports
at once, but to also be able to document _why_ those exclusions are
being performed.

Thus, allow for an `--exclude-file` to be provided.  One entry per line,
and which can be documented with `#` style comments.