add support for output file in JSON format
Having the audit result in JSON format makes it easier to integrate cpan-audit in CI/CD pipelines or importing
the results in other apps used to keep track of security issues (e.g. DefectDojo).

It doesn't print the JSON on the STDOUT as other log messages are already printed on STDOUT.