Allow for JSON output (#31)
* add support for output file in JSON format

Having the audit result in JSON format makes it easier to integrate cpan-audit in CI/CD pipelines or importing
the results in other apps used to keep track of security issues (e.g. DefectDojo).

It doesn't print the JSON on the STDOUT as other log messages are already printed on STDOUT.

* add test for new --json feature

* add new --json option to documentation and examples

* update meta files

* 'Fix' deps test.

exclude CVE-2011-4116 explicitly. It's a known issue in File::Temp wrt symlinks.
It should be safe to use the module the way we use it though.

* Ensure TestCommand uses the modules in the dist

* Move almost all output to cpan-audit

* Don't output verbose messages under --quiet

* Update reports database

Co-authored-by: reneeb <info@perl-services.de>